LayoutTests/css3/calc/margin-expected.txt

PASS computedMarginLeft("simple-all") is "25px"
PASS computedMarginTop("simple-all") is "25px"
PASS computedMarginRight("simple-all") is "25px"
PASS computedMarginBottom("simple-all") is "25px"
PASS computedMarginLeft("simple-left") is "25px"
PASS computedMarginTop("simple-left") is "0px"
PASS computedMarginRight("simple-left") is "0px"
PASS computedMarginBottom("simple-left") is "0px"
PASS computedMarginLeft("simple-right") is "0px"
PASS computedMarginTop("simple-right") is "0px"
PASS computedMarginRight("simple-right") is "25px"
PASS computedMarginBottom("simple-right") is "0px"
PASS computedMarginLeft("simple-top") is "0px"
PASS computedMarginTop("simple-top") is "25px"
PASS computedMarginRight("simple-top") is "0px"
PASS computedMarginBottom("simple-top") is "0px"
PASS computedMarginLeft("simple-bottom") is "0px"
PASS computedMarginTop("simple-bottom") is "0px"
PASS computedMarginRight("simple-bottom") is "0px"
PASS computedMarginBottom("simple-bottom") is "25px"
PASS computedMarginLeft("percent-all") is "25px"
PASS computedMarginTop("percent-all") is "25px"
PASS computedMarginRight("percent-all") is "25px"
PASS computedMarginBottom("percent-all") is "25px"
PASS computedMarginLeft("percent-left") is "25px"
PASS computedMarginTop("percent-left") is "0px"
PASS computedMarginRight("percent-left") is "0px"
PASS computedMarginBottom("percent-left") is "0px"
PASS computedMarginLeft("percent-right") is "0px"
PASS computedMarginTop("percent-right") is "0px"
PASS computedMarginRight("percent-right") is "25px"
PASS computedMarginBottom("percent-right") is "0px"
PASS computedMarginLeft("percent-top") is "0px"
PASS computedMarginTop("percent-top") is "25px"
PASS computedMarginRight("percent-top") is "0px"
PASS computedMarginBottom("percent-top") is "0px"
PASS computedMarginLeft("percent-bottom") is "0px"
PASS computedMarginTop("percent-bottom") is "0px"
PASS computedMarginRight("percent-bottom") is "0px"
PASS computedMarginBottom("percent-bottom") is "25px"
PASS successfullyParsed is true

TEST COMPLETE
